Created `big_commit.sh`
Created big_commit.sh. Script takes in a tar, unzips and uncompresses
it, and then deletes all files with "DELETE ME!" in them. It then
repackages them in another tgz file following `cleaned_..` where the ..
is the name of the original tgz file. It is then placed in the same directory as the shell script itself.